Changed getStringArgument to use wchar_t instead of char, after a report of Alert() and related string functions not working I am sure this worked with char at some point. But indeed my current brokers MT4 client is providing a wide character buffer for the strings... I wonder if this was a mistake previously, or if this is different for different Broker clients, or perhaps older builds of MT4.

+++ PyMT4/pymt4_common_iosession.cpp
@@ -94,7 +94,7 @@ void IOSessionCommon::headerHandler(const boost::system::error_code& error)
 
 		Buffer::iterator readBufferIter = m_readBuffer.begin();
 		MessageTypeIdentifier			messageTypeIdentifier;
-		size_t							messageContentSize;
+		uint32_t							messageContentSize;
 
 		Serializer<MessageTypeIdentifier>::deserializeItem(&messageTypeIdentifier,&readBufferIter);
 		Serializer<uint32_t>::deserializeItem(&messageContentSize,&readBufferIter);
